src/yarn.rs: move the async yarn-serve server back here
