+++ src/serve.rs
@@ -17,8 +17,6 @@ use std::{convert::Infallible, path::Path, sync::Arc, time::Duration};
 use tectonic_errors::prelude::*;
 use tectonic_status_base::StatusBackend;
 use tokio::{
-    io::{AsyncBufReadExt, BufReader},
-    process::{Child, Command},
     signal::unix::{signal, SignalKind},
     sync::{mpsc, oneshot, Mutex},
 };
@@ -28,7 +26,10 @@ use warp::{
     Filter, Rejection, Reply,
 };
 
-use crate::messages::{Message, MessageBus, ToolOutputStream, YarnOutputMessage};
+use crate::{
+    messages::{Message, MessageBus},
+    yarn::YarnServer,
+};
 
 /// The serve operation.
 #[derive(Args, Debug)]
@@ -38,7 +39,7 @@ pub struct ServeArgs {
 }
 
 /// A message to be delivered to the main "serve" thread.
-enum ServeCommand {
+pub enum ServeCommand {
     /// Rebuild the document.
     Build,
 
@@ -369,129 +370,6 @@ impl DebounceEventHandler for Watcher {
             futures::executor::block_on(async {
                 self.command_tx.send(ServeCommand::Build).await.unwrap()
             });
-        }
-    }
-}
-
-// The `yarn serve` task
-
-struct YarnServer {
-    child: Child,
-    quit_rx: oneshot::Receiver<()>,
-    command_tx: mpsc::Sender<ServeCommand>,
-    clients: WarpClientCollection,
-}
-
-impl YarnServer {
-    fn new(
-        port: u16,
-        quit_rx: oneshot::Receiver<()>,
-        command_tx: mpsc::Sender<ServeCommand>,
-        clients: WarpClientCollection,
-    ) -> Result<Self> {
-        let mut cmd = Command::new("yarn");
-        cmd.arg("serve")
-            .arg(format!("--port={port}"))
-            .arg("--watch-for-stdin")
-            .stdin(std::process::Stdio::piped())
-            .stdout(std::process::Stdio::piped())
-            .stderr(std::process::Stdio::piped());
-
-        let child = atry!(
-            cmd.spawn();
-            ["failed to launch `yarn serve` process"]
-        );
-
-        Ok(YarnServer {
-            child,
-            quit_rx,
-            command_tx,
-            clients,
-        })
-    }
-
-    async fn serve(mut self) {
-        let stdin = self.child.stdin.take().expect("failed to open child stdin");
-
-        let stdout = self
-            .child
-            .stdout
-            .take()
-            .expect("failed to open child stdout");
-        let mut stdout_lines = BufReader::new(stdout).lines();
-
-        let stderr = self
-            .child
-            .stderr
-            .take()
-            .expect("failed to open child stderr");
-        let mut stderr_lines = BufReader::new(stderr).lines();
-
-        loop {
-            tokio::select! {
-                _ = &mut self.quit_rx => {
-                    break;
-                },
-
-                line = stdout_lines.next_line() => {
-                    match line {
-                        Ok(Some(line)) => {
-                            self.clients.post(&Message::YarnOutput(YarnOutputMessage {
-                                stream: ToolOutputStream::Stdout,
-                                lines: vec![line],
-                            })).await;
-                        }
-
-                        Err(e) => {
-                            eprintln!("error: failed to read `yarn serve` output: {e}");
-                        }
-
-                        _ => {}
-                    }
-                }
-
-                line = stderr_lines.next_line() => {
-                    match line {
-                        Ok(Some(line)) => {
-                            self.clients.post(&Message::YarnOutput(YarnOutputMessage {
-                                stream: ToolOutputStream::Stderr,
-                                lines: vec![line],
-                            })).await;
-                        }
-
-                        Err(e) => {
-                            eprintln!("error: failed to read `yarn serve` output: {e}");
-                        }
-
-                        _ => {}
-                    }
-                }
-
-                status = self.child.wait() => {
-                    let msg = match status {
-                        Ok(s) => format!("`yarn serve` subprocess exited early, {s}"),
-                        Err(e) => format!("`yarn serve` subprocess exited early and failed to get outcome: {e}"),
-                    };
-
-                    let cmd = ServeCommand::Quit(Err(anyhow!(msg.clone())));
-
-                    if let Err(e) = self.command_tx.send(cmd).await {
-                        eprintln!("error: {msg}");
-                        eprintln!("  ... furthermore, the yarn task failed to notify main task to exit: {e:?}");
-                    }
-
-                    break;
-                }
-            }
-        }
-
-        // Close stdin, informing the process to exit (thanks to `--watch-for-stdin`)
-        std::mem::drop(stdin);
-
-        let status = self.child.wait().await;
-
-        if status.is_err() {
-            eprintln!("error: `yarn serve` subprocess exited with an error: {status:?}");
         }
     }
 }
